File naming correction and moved existing tests
