+++ torchmetrics/__init__.py
@@ -61,4 +61,5 @@ from torchmetrics.retrieval import (  # noqa: F401 E402
     RetrievalPrecision,
     RetrievalRecall,
 )
+from torchmetrics.text import BLEUScore  # noqa: F401 E402
 from torchmetrics.wrappers import BootStrapper  # noqa: F401 E402
